fix: fix how the auto setting is applied
When auto scaling is selected, clear local storage scaling value
by setting it to undefined in order to use the calculated auto value

refs: CDS-78
