+++ src/settings/components/utils.ts
@@ -7,6 +7,7 @@
 import moment from 'moment';
 import { TFunction } from 'react-i18next';
 import { AccountSettings } from '../../../types';
+import { BASE_FONT_SIZE, SCALING_LIMIT, SCALING_OPTIONS } from '../../constants';
 
 export const ItemsSendAutoReplies = (t: TFunction): any => [
 	{
@@ -1287,3 +1288,17 @@ export const timeZoneList = (t: TFunction): Array<TimeZoneDescriptor> => [
 		})
 	}
 ];
+
+export const getAutoScalingFontSize = (): number => {
+	if (
+		window.screen.width <= SCALING_LIMIT.WIDTH &&
+		window.screen.height <= SCALING_LIMIT.HEIGHT &&
+		window.devicePixelRatio >= SCALING_LIMIT.DPR
+	) {
+		const baseFontIndex = SCALING_OPTIONS.findIndex((option) => option.value === BASE_FONT_SIZE);
+		if (baseFontIndex > 0) {
+			return SCALING_OPTIONS[baseFontIndex - 1].value;
+		}
+	}
+	return BASE_FONT_SIZE;
+};
